Stop using isOnFrontPage prop. - Move associated styling to src/app/pages/FrontPage/index.jsx.
