+++ src/app/containers/MostRead/index.jsx
@@ -14,7 +14,6 @@ const MostReadContainer = ({
   mostReadEndpointOverride,
   maxTwoColumns,
   constrainMaxWidth,
-  isOnFrontPage,
 }) => {
   const { variant, isAmp } = useContext(RequestContext);
   const {
@@ -38,7 +37,6 @@ const MostReadContainer = ({
       endpoint={endpoint}
       constrainMaxWidth={constrainMaxWidth}
       maxTwoColumns={maxTwoColumns}
-      isOnFrontPage={isOnFrontPage}
     />
   );
 };
@@ -47,14 +45,12 @@ MostReadContainer.propTypes = {
   mostReadEndpointOverride: string,
   constrainMaxWidth: bool,
   maxTwoColumns: bool,
-  isOnFrontPage: bool,
 };
 
 MostReadContainer.defaultProps = {
   mostReadEndpointOverride: null,
   constrainMaxWidth: false,
   maxTwoColumns: false,
-  isOnFrontPage: false,
 };
 
 export default MostReadContainer;
